fix: https-proxy volume, user, working_dir removed, updated https-portal image version
Currently https-portal only works with the root user but this wouldn't be an issue since we don't mount any volumes any more.
symbolbootstrap stop command will actually stop containers and run command will start, unless there is a version upgrade(which will trigger a new container creation)
it wouldn't need to re-sign the certificates. So most of the time it will be reusing the same certificates.